pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mrv-lib"
version = "0.2.0"
authors = [
  { name = "Kai Zheng", email = "kai.zheng@mrv-lib.org" },
]
description = "Model Risk Validator for Model Risk Governance"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Topic :: Office/Business :: Financial",
]
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
]

[project.optional-dependencies]
ib = ["ib_insync"]
validator = ["scikit-learn", "hmmlearn", "scipy", "matplotlib"]
dev = ["pytest", "pytest-cov", "black", "ruff", "scikit-learn", "scipy", "matplotlib"]
all = ["ib_insync", "scikit-learn", "hmmlearn", "scipy", "matplotlib"]

[project.urls]
Homepage = "https://mrv-lib.org"
"Bug Tracker" = "https://github.com/modelguard-lab/mrv-lib/issues"
Repository = "https://github.com/modelguard-lab/mrv-lib"

# CLI entry: use `python run.py` directly (run.py is not a package module)
# [project.scripts]
# mrv = "mrv.cli:main"  # uncomment when cli.py is restored

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mrv = ["../../config.yaml"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning:sklearn",
]
